+++ src/Header/CacheControl.php
@@ -3,6 +3,8 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
+use Innmind\Http\Header;
+
 /**
  * @psalm-immutable
  */