+++ src/Header/Cookie.php
@@ -3,6 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
+use Innmind\Http\Header;
 use Innmind\Immutable\Map;
 
 /**